Remove a parasite (test)... :'(
